+++ PokemonEncyclopedia/PokemonEncyclopedia/FeedCache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -69,50 +69,3 @@ public class CoreDataFeedStore: FeedStore {
 	}
 	
 }
-
-
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-	@NSManaged var timestamp: Date
-	@NSManaged var feed: NSOrderedSet
-	
-	static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-		try find(in: context).map(context.delete)
-		return ManagedCache(context: context)
-	}
-	
-	var localFeed:[LocalFeedImage] {
-		return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
-	}
-	
-	static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-		let request = NSFetchRequest<ManagedCache>(entityName: entity().name!)
-		request.returnsObjectsAsFaults = false
-		return try context.fetch(request).first
-	}
-	
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-	@NSManaged var id: UUID?
-	@NSManaged var name: String?
-	@NSManaged var url: URL
-	@NSManaged var cache: ManagedCache
-	
-	static func image(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-		return NSOrderedSet(array: localFeed.map { local in
-			let managed = ManagedFeedImage(context: context)
-			managed.id = local.id
-			managed.name = local.name
-			managed.url = local.url
-			return managed
-		})
-	}
-	
-	var local: LocalFeedImage {
-		return LocalFeedImage(id: id, name: name, url: url)
-	}
-	
-}
